--- package.json
@@ -1,32 +1,0 @@
-{
-  "name": "react-typescript",
-  "version": "0.1.0",
-  "private": true,
-  "main": "src/index.tsx",
-  "dependencies": {
-    "prop-types": "15.7.2",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-router-dom": "5.0.1",
-    "react-scripts-ts": "^2.13.0",
-    "styled-components": "4.3.2"
-  },
-  "scripts": {
-    "start": "react-scripts-ts start",
-    "build": "react-scripts-ts build",
-    "test": "react-scripts-ts test --env=jsdom",
-    "eject": "react-scripts-ts eject"
-  },
-  "devDependencies": {
-    "@types/jest": "^22.1.3",
-    "@types/node": "^9.4.6",
-    "@types/react": "16.0.38",
-    "@types/react-dom": "16.0.4",
-    "typescript": "^2.7.2"
-  },
-  "keywords": [
-    "typescript",
-    "react"
-  ],
-  "description": "Adding types to React!"
-}
