--- src/index.css
@@ -1,13 +1,0 @@
-:root {
-  font-size: 10px;
-}
-body {
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-}
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: content-box;
-}
